requirements.txt
Django
python-decouple
twilio
phonenumbers